RC-22: moved to spring 3.1.5

+++ src/main/resources/db/migration/V1_02__create_contacts_table.sql
@@ -2,7 +2,7 @@ create table realtors_contacts
 (
     id         bigserial     not null primary key,
     contact    varchar(2048) not null,
-    type_id    smallint      not null,
+    type_id    integer       not null,
     realtor_id bigint        not null
         constraint fk_to_realtor_id
             references realtors_info
